Reformatted sync read function
Reformatted the sync read function to be similiar in nature to the
function found in the c rtl-sdr library, having blocks that are then
appended to a total rather than one massive read.

I also fixed a typo that I had in my python function that kept only half
the samples I wanted to display, now it instead keeps the whole amount.

Attached pic is of 99.5 megahertz band with no bandwidth limit imposed,
which I believe will make the bandwidth equal to the sample rate or half
the sample rate, at around 2 megahertz. The spectrogram I got clearly
looks like what is found at that band, with 5 wideband signals in site,
and testing with GQRX, a c++ spectrum analyzer I found the same 5
bands.

I still need to play around with the gain or the bandwidth however,
becasue if I were to run the same test using python functions to
generate the IQ data, the inside signals which are a very weak FM
signal and what appears to be the same signal of the center FM frequency
but much weaker (I'm sure there is some sort of DSP explaination for why
that it shows up), dont show up in the python version. I will play
around with the automatic gain and the passband to try and figure out
why this happens