Added a few new C calls, did some testing on fm signals
Added C calls for setting the tuner BW and for setting Automatic gain
control to be on. Not sure how useful these are as BW control is
probably easier to handle via sampling rate and Digital decimation and
AGC I'm assuming is enabled by default. I should read through the python
implementation of RTL_SDR to see what they intialize things to be.

I did some more testing on fm signals and I found that I can get a
decent spectrum of an FM signal and filter out all the surrounding
signals, but I am having trouble demodulating said signal to be sound.
I'm gonna analize the signal at the different stages of
demodulation/processing to see where exactly the problem is, but I think
it might be something to do with the gain not being high enough

I also want to make a scripting folder for testing the spectrum and the
fm_demodulations