Added functions for getting magnitude for IQ data
Added 2 functions for getting the magnitude from IQ data, get_mag and
get_mag_quick. get mag uses the tradional a^2 + b^ = c^2, but if that
proves to be too computationally slow I have get_mag_quick, which uses a
little dsp trick, where c = alpha*max(|a|, |b|) + beta*min(|a|, |b|),
where alpha and beta can be chosen based on optimizing for rms
calculation errors, peak calculation errors, or simplicity. For now I'm
using literally 1 and 1/4, with a little bit of jank becuase I only have
integer values not floats for now. It doesn't matter a whole lot
globally becuase Mode-S messages are encoded using PPM, so the only
error that can occur is if a 1 is mistaken for a 0 so we have a little
bit of room for error (hopefully).

To implement this I had to change how I formatted a couple of things in
IQ data. There is room for improvement 100% but for now I just want to
try and get a base implementation then work on optimization and making
the code look cool.