Added methods to check for preamble in ads_b
Added a method that returns the total number of preambles found in
sequence of IQ data. Still need to test it to see how well it worked.
Every part of it (1 vs 0 algorthim, testing alg, iteration method, etc)
can proably be significantly improved but for now I just want to get a
baseline to work with