Wrote functions for doing crc checksum calculation
Was diving deep into rust documentation to do a cool looking iterator to
help out with my crc checksum. Ended up just using a for loop lol. Still
need to test to see if I implemented the algorithm correctly, but the
idea is that it uses a 24 digit shift register that xors with the current value
in the register if the value just pushed out is a 1.