Wrote a simple crc test that doesn't rely on anything else
Wrote a simple crc tester in python that I used to generate some crc
encoded polynomials that I passed to my function. All of them returned
true which means that the "problem" right now in the pipeline isn't with
the crc.

The problem has to be one of or a combination of these 5:
1. The preamble detection is wrong, so the polynomials passed in aren't
correct in the first place
2. The formatting between the preamble detection and the crc detection
has a simple error (even just 1 off would mess up everything),
preventing it from ever working
3. The signal is so weak that it cannot work properly without 1 or 2 bit
correction (unlikely)
4. The signal requires some form a phase correction to be properly
detected
5. The differential decoding for the signal is handled incorrectly