Small changes made that hopefully make crc pass
So I made some small changes that I am hoping will allow the system to
pass its crc checks. For one I figured out the whole thing with the
ranges. The way it works is that it goes inclusive,exclusive so for
getting 16 samples i do want to do i..i+16, which was confusing me at
first.

I also found that I was skipping 224 ahead in the case of preambles that
had a preamble match but failed the crc check rather than skipping 1
ahead. ofc this will make it now run slower so I set the number of
samples to capture to be lower. When I get a good signal I will test to
see if my changes hopefully worked :-)