Did unit testing of all of my signal processing functions
Did some unit testing of all of my signal processing functions and at
the basic level they all passed. This is somewhat helpful for finding
where I am going wrong because it means that I at least didn't mess up
at the basic implementations of my stuff. I also sorta implemented some
psuedo phase correction that goes into effect if the basic crc check
fails. What this all tells me about the error is that one or more of the
following is true:

1. Somewhere you messed up with your rtl-sdr IQ data capturing
implementation: Which is kinda hard to check but at the very least I can
see if I can get it to decode fm radio, which is always an effective
test.
2. The format you are trying to decode (DF 17) is very rare and you
should also implement other decoding methods to get anything worthwhile: I can check this by having
it check the first 5 bits before it decodes the whole message.
3. dump1090 can decode signals but you cannot because it has a lot of
jank to it that allows it to boost the signal a whole bunch: This one is
the most hard to test but could be done by trying to get a better
reception on the signal either outside or with a better antenna.
4. The method of block size + n samples allows for all information to
slip through and you need a better method of capturing data such as
async reads to get any results: Not sure how to really test this other
then just playing around with block size and number of samples captured.
At some point I will re-write sdr_reader to be more reusable and not
print out as much information.